aplicação de CQRS e criacao de controllers

+++ CleanArch.Domain/Entities/Product.cs
@@ -25,7 +25,7 @@ namespace CleanArch.Domain.Entities
             ValidateDomain(name, description, price, stock, image);
         }
 
-        public void Update(string name, string description, decimal price, int stock, string image, int categoryId)
+        public void Update(string name, string description, decimal price, int stock, string image, long categoryId)
         {
             ValidateDomain(name, description, price, stock, image);
             CategoryId = categoryId;
